+++ marketplace/backend/dapp/items/contracts/Membership.sol
@@ -12,11 +12,13 @@ contract Membership is SemiFungible {
         string _description,
         string[] _images,
         string[] _files,
+        string[] _fileNames,
         uint _createdDate,
         uint _quantity,
         uint _expirationPeriodInMonths,
-        AssetStatus _status
-    ) public SemiFungible(_name, _description, _images, _files, _createdDate, _quantity, _status) {
+        AssetStatus _status,
+        address _redemptionService
+    ) public SemiFungible(_name, _description, _images, _files, _fileNames, _createdDate, _quantity, _status, _redemptionService) {
         expirationPeriodInMonths = _expirationPeriodInMonths;
         expirationDate = block.timestamp + (expirationPeriodInMonths*2592000);
     }
@@ -27,10 +29,12 @@ contract Membership is SemiFungible {
             description,
             images,
             files,
+            fileNames,
             createdDate,
             _quantity,
             expirationPeriodInMonths,
-            status
+            status,
+            address(redemptionService)
         );
         return UTXO(address(newAsset));
     }