+++ marketplace/ui/src/components/MarketPlace/NewTrendingCard.js
@@ -1,28 +1,32 @@
-import { HeartFilled, HeartTwoTone } from '@ant-design/icons'
-import { useAuthenticateState } from "../../contexts/authentication";
 import React, { useState, useEffect } from "react";
-import {
-    Typography,
-    Button,
-    InputNumber,
-    Tooltip
-} from "antd";
+import { Typography, Button, InputNumber, Tooltip } from "antd";
 import { useLocation, useNavigate } from "react-router-dom";
-import routes from "../../helpers/routes";
+import { HeartFilled, HeartTwoTone } from '@ant-design/icons'
 import TagManager from "react-gtm-module";
-import { Images } from '../../images';
-import images_placeholder from "../../images/resources/image_placeholder.png"
-import { SEO } from '../../helpers/seoConstant';
 import DOMPurify from 'dompurify';
+// State
+import { useAuthenticateState } from "../../contexts/authentication";
+// Assets
+import images_placeholder from "../../images/resources/image_placeholder.png";
+import { Images } from '../../images';
+// other
+import { STRATS_CONVERSION } from '../../helpers/constants';
 import { setCookie } from "../../helpers/cookie";
+import { SEO } from '../../helpers/seoConstant';
+import routes from "../../helpers/routes";
 import LoginModal from './LoginModal';
 
 const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, contextHolder, isUserProfile = false }) => {
+    const navigate = useNavigate();
+    const location = useLocation();
     const { Text } = Typography;
-    const [quantity, setQuantity] = useState(1)
-    const [isModalVisible, setIsModalVisible] = useState(false);
 
-    let { hasChecked, isAuthenticated, loginUrl, user } = useAuthenticateState();
+    const { hasChecked, isAuthenticated, loginUrl, user } = useAuthenticateState();
+
+    const [isModalVisible, setIsModalVisible] = useState(false);
+    const [isWishlisted, setIsWishlisted] = useState(false);
+    const [quantity, setQuantity] = useState(1);
+
     const ownerSameAsUser = () => {
         if (user?.commonName === topSellingProduct?.ownerCommonName) {
             return true;
@@ -30,19 +34,13 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
         return false;
     }
 
-    // For Wishlist Icon Rendering
-    const [isWishlisted, setIsWishlisted] = useState(false);
-    const shouldShowWishlistIcon = isAuthenticated && user && !ownerSameAsUser();
-
     const naviroute = routes.MarketplaceProductDetail.url;
     const isAvailableForSale = (!topSellingProduct.price || topSellingProduct.saleQuantity === 0)
-    const navigate = useNavigate();
-    const location = useLocation();
 
     const queryParams = new URLSearchParams(location.search);
     const categoryQueryValue = queryParams.get('category');
-    const categoryQueryValueArr = categoryQueryValue ? categoryQueryValue.split(',') : []
-    const imgMeta = categoryQueryValueArr.length === 1 ? categoryQueryValueArr[0] : SEO.IMAGE_META
+    const categoryQueryValueArr = categoryQueryValue ? categoryQueryValue.split(',') : [];
+    const imgMeta = categoryQueryValueArr.length === 1 ? categoryQueryValueArr[0] : SEO.IMAGE_META;
 
     const sanitizedDescription = DOMPurify.sanitize(topSellingProduct?.description || "N/A");
     const customStyle = {
@@ -60,7 +58,7 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
     }, [topSellingProduct]);
 
     const toggleWishlist = () => {
-        if (!shouldShowWishlistIcon) {
+        if (!isAuthenticated || !user) {
             setIsModalVisible(true);
         } else {
             const wishList = JSON.parse(localStorage.getItem('wishList')) || [];
@@ -90,13 +88,16 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
         setIsModalVisible(false);
     };
 
+
     return (
         <>
-            <div id='productCard' className={`relative trending_cards_container_card bg-white p-3 ${parent == 'Marketplace' ? 'min-w-[300px] w-auto' : 'min-w-[230px]'}  min-w-[320px] md:min-w-[300px] rounded-md flex flex-col gap-2 md:gap-3 shadow-card_shadow h-max`}>
+            <div id='productCard' className={`relative trending_cards_container_card bg-white p-3 ${parent === 'Marketplace' ? 'min-w-[300px] w-auto' : 'min-w-[230px]'}  min-w-[320px] md:min-w-[300px] rounded-md flex flex-col gap-2 md:gap-3 shadow-card_shadow h-max`}>
                 {contextHolder}
-                <div onClick={toggleWishlist} className="absolute top-2 right-2 cursor-pointer hover:scale-110 transition-transform duration-200">
-                    {isWishlisted ? <HeartFilled style={{ fontSize: "20px", color: "#A15E49" }} /> : <HeartTwoTone style={{ fontSize: "20px" }} twoToneColor="#A15E49" />}
-                </div>
+                {!ownerSameAsUser() &&
+                    <div onClick={toggleWishlist} className="absolute top-2 right-2 cursor-pointer hover:scale-110 transition-transform duration-200">
+                        {isWishlisted ? <HeartFilled style={{ fontSize: "20px", color: "#A15E49" }} /> : <HeartTwoTone style={{ fontSize: "20px" }} twoToneColor="#A15E49" />}
+                    </div>
+                }
                 <a
                     href={`${naviroute.replace(":address", topSellingProduct.address).replace(":name", topSellingProduct.name)}`}
                     onClick={(e) => {
@@ -112,7 +113,7 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
                 >
                     <img
                         className='md:h-[200px] md:w-[40vw] h-[150px] w-full object-contain rounded-md cursor-pointer mb-2'
-                        src={topSellingProduct.images ? topSellingProduct?.images[0] : images_placeholder}
+                        src={topSellingProduct["BlockApps-Mercata-Asset-images"]?.length > 0 ? topSellingProduct["BlockApps-Mercata-Asset-images"][0].value : images_placeholder}
                         alt={imgMeta} title={imgMeta}
                     />
                     <div className='flex justify-between items-center'>
@@ -130,7 +131,12 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
                     </div>
                 </a>
                 <div style={{ display: 'flex', justifyContent: 'space-between' }}>
-                    {topSellingProduct?.price && <Typography className='font-normal text-black'>{`$ ${topSellingProduct?.price}`}</Typography>}
+                    {topSellingProduct?.price &&
+                        <Typography className="font-semibold">
+                        {`$${topSellingProduct?.price} `} <span className="font-normal text-xs mr-2 text-primary"><b> {`(${topSellingProduct?.price * STRATS_CONVERSION} STRATS)`} </b></span>
+                    </Typography>
+                    
+                    }
                     {isAvailableForSale && <Text type="danger" strong> Sold Out </Text>}
                     {topSellingProduct?.contract_name.toLowerCase().includes("clothing") && (
                         <Typography className='font-normal text-black'>Size: {topSellingProduct?.data?.size ? topSellingProduct?.data?.size : "N/A"}</Typography>
@@ -146,7 +152,7 @@ const NewTrendingCard = ({ topSellingProduct, addItemToCart, parent = "", api, c
                     <Typography>Quantity:</Typography>
                     <div className='flex gap-3 p-1 bg-white'>
                         <Typography className={`px-2 bg-[#EEEFFA] rounded-sm ${quantity === 1 ? 'cursor-not-allowed opacity-50' : 'cursor-pointer'}`} onClick={() => {
-                            setQuantity(quantity == 1 ? quantity : quantity - 1)
+                            setQuantity(quantity === 1 ? quantity : quantity - 1)
                         }}>
                             -
                         </Typography>