+++ marketplace/ui/src/components/Order/SoldOrderDetails.js
@@ -45,7 +45,6 @@ const SoldOrderDetails = ({ user, users }) => {
   const [selectedDate, setSelectedDate] = useState("");
   const [status, setStatus] = useState(getStatus(1));
   const [paid, setPaid] = useState("Processing");
-  const [isLoadingPaymentStatus, setisLoadingPaymentStatus] = useState(false)
   const [comment, setComment] = useState("");
   const { TextArea } = Input;
   const [api, contextHolder] = notification.useNotification();
@@ -67,27 +66,35 @@ const SoldOrderDetails = ({ user, users }) => {
 
   useEffect(() => {
     if (orderDetails) {
-      setStatus(getStatus(parseInt(orderDetails.order.status)));
-      setComment(orderDetails.order.comments);
-      // Fulfillment date is sometimes coming in as 0. a unix of 0 sets the date to 1969. So we need to check for 0 and null, I added undefined just in case too. 
-      if (orderDetails.order.fulfillmentDate === 0 || orderDetails.order.fulfillmentDate === null || orderDetails.order.fulfillmentDate === undefined) {
-        setSelectedDate(null);
-      } else {
-        setSelectedDate(dayjs.unix(orderDetails.order.fulfillmentDate));
+      const statusInt = parseInt(orderDetails.order.status);
+      setStatus(getStatus(statusInt));
+      if (statusInt === 3) {
+        setPaid("Paid");
+      } else if (statusInt === 4) {
+        setPaid("Canceled");
       }
+      setComment(orderDetails.order.comments);
+      // Order Close Date is represented by block_timestamp when the Order Status is 3(CLOSED) or 4(CANCELED). This is consistent across legacy orders and new orders as there wouldn't be updates/methods invoked when the Order Status reaches Closed.
+      if (parseInt(orderDetails.order.status) === 3 || parseInt(orderDetails.order.status) === 4) {
+        const formattedDate = dayjs(orderDetails.order.block_timestamp);
+        setSelectedDate(formattedDate);
+      } else {
+        setSelectedDate(null);
+      } 
 
       let items = [];
+      const orderQuantities = orderDetails.order.quantities ? orderDetails.order.quantities : orderDetails.order["BlockApps-Mercata-Order-quantities"].map(item => item.value);
       orderDetails.assets.forEach((prod, index) => {
         items.push({
           address: prod.address,
           chainId: prod.chainId,
           key: prod.address,
-          productImage: prod.images && prod.images.length > 0 ? prod.images[0] : image_placeholder,
+          productImage: prod["BlockApps-Mercata-Asset-images"].length > 0 ? prod["BlockApps-Mercata-Asset-images"][0].value : image_placeholder,
           productName: prod,
           name: prod.name,
           unitPrice: prod.price,
-          quantity: parseInt(orderDetails.order.quantities[index]),
-          amount: prod.price * parseInt(orderDetails.order.quantities[index]),
+          quantity: parseInt(orderQuantities[index]),
+          amount: prod.price * parseInt(orderQuantities[index]),
           serialNumber: prod,
           tax: prod.tax ? prod.tax : 0,
         });
@@ -107,81 +114,8 @@ const SoldOrderDetails = ({ user, users }) => {
   }, [Id, dispatch, status]);
 
   const getData = async () => {
-    const data = await actions.fetchOrderDetails(dispatch, Id);
-    if (data != null) {
-      getPaymentStatus(data.order.paymentSessionId, data.order.sellersCommonName);
-    }
+    await actions.fetchOrderDetails(dispatch, Id);
   };
-
-  const validatePayment = async (paymentSessionId) => {
-    if (!paymentSessionId || !orderDetails) return;
-
-    const currentStatus = getStatus(parseInt(orderDetails.order.status));
-    const isPending = currentStatus === getStatusByName("Payment Pending")
-    const isCanceled = currentStatus === getStatusByName("Canceled");
-
-    if (isCanceled) {
-      setPaid("Payment Failed");
-      setComment(orderDetails.order.comments);
-    }
-    if (isPending) {
-      try {
-        const response = await fetch(
-          `${apiUrl}/order/payment/intent/${paymentSessionId}/${orderDetails.order.sellersCommonName}`,
-          { method: HTTP_METHODS.GET }
-        );
-        const intentBody = await response.json();
-        const paymentErrorAndRequiresMethod = intentBody.data.last_payment_error?.message && intentBody.data.status === 'requires_payment_method';
-
-        if (paymentErrorAndRequiresMethod && !isCanceled) {
-          setisLoadingPaymentStatus(true)
-          const body = {
-            saleOrderAddress: orderDetails.order.address,
-            comments: encodeURIComponent('Stripe: ' + intentBody.data.last_payment_error.message),
-          };
-          //Update Order Details and change the Order Status to 'Canceled' from 'Payment Pending'
-          let isDone = await actions.cancelSale(dispatch, body);
-          setComment(`Stripe: ${orderDetails.order.comments}`);
-          if (isDone) {
-            setStatus("Canceled");
-            setPaid("Payment Failed");
-            await actions.fetchOrderDetails(dispatch, Id);
-            setisLoadingPaymentStatus(false);
-          }
-        }
-
-      } catch (err) {
-        console.error(`Error: ${err}`);
-      }
-    }
-  };
-
-  const getPaymentStatus = async (paymentSessionId, sellersCommonName) => {
-    if (!paymentSessionId) return;
-
-    setisLoadingPaymentStatus(true);
-    try {
-      const response = await fetch(
-        `${apiUrl}/order/payment/session/${paymentSessionId}/${sellersCommonName}`,
-        { method: HTTP_METHODS.GET }
-      );
-
-      const body = await response.json();
-      if (response.status === RestStatus.OK) {
-        if (body.data["payment_status"] === "paid") {
-          setPaid("Paid");
-        } else {
-          await validatePayment(paymentSessionId);
-        }
-      }
-    } catch (err) {
-      console.error(`Error: ${err}`);
-    } finally {
-      setisLoadingPaymentStatus(false);
-    }
-  };
-
-
 
   const details = orderDetails;
   const audits = ordersAudit;
@@ -296,6 +230,10 @@ const SoldOrderDetails = ({ user, users }) => {
         textClass: "bg-[#FFF0F0]",
         bgClass: "bg-[#FF0000]"
       },
+      ["Canceled"]: {
+        textClass: "bg-[#FFF0F0]",
+        bgClass: "bg-[#FF0000]"
+      }
     };
 
     const { textClass, bgClass } = statusClasses[status] || { textClass: "bg-[#FFF6EC]", bgClass: "bg-[#119B2D]" };
@@ -385,14 +323,6 @@ const SoldOrderDetails = ({ user, users }) => {
   return (
     <div>
       {contextHolder}
-      {details === null || isorderDetailsLoading || isLoadingPaymentStatus ? (
-        <div className="h-screen flex justify-center items-center">
-          <Spin
-            spinning={isorderDetailsLoading || isLoadingPaymentStatus}
-            size="large"
-          />
-        </div>
-      ) : (
         <div>
           <Breadcrumb className="text-sm ml-4 md:ml-20  mt-0 md:mt-5 mb-2">
             <Breadcrumb.Item href="" onClick={e => e.preventDefault()}>
@@ -406,7 +336,7 @@ const SoldOrderDetails = ({ user, users }) => {
               </div>
             </Breadcrumb.Item>
             <Breadcrumb.Item className="text-sm text-[#202020] font-medium">
-              {details.order.orderId}
+              {`${details?.order?.orderId || ''}`.substring(0,6)}
             </Breadcrumb.Item>
           </Breadcrumb>
 
@@ -421,6 +351,14 @@ const SoldOrderDetails = ({ user, users }) => {
                 children:
                   <div className="mb-10">
                     <Button type="ghost" onClick={() => onChange('Sold')} className="cursor-pointer px-2 flex md:hidden items-center gap-2 text-xs font-semibold"><LeftArrow /> Back</Button>
+                    {details === null || isorderDetailsLoading ? (
+                      <div className="h-screen flex justify-center items-center">
+                        <Spin
+                          spinning={isorderDetailsLoading}
+                          size="large"
+                        />
+                      </div>
+                    ) : (
                     <Card className="md:p-2 mb-4 md:mb-14 md:shadow-card_shadow order_detail_card">
                       <div className="flex flex-col md:flex-row md:justify-between">
                         <div className="flex flex-col">
@@ -459,7 +397,7 @@ const SoldOrderDetails = ({ user, users }) => {
                         </Button>
                       </div>
                       <Row className="hidden md:flex my-6 justify-between bg-[#F6F6F6] p-4 pb-2 rounded">
-                        <OrderData title="Order Number" value={`#${details.order.orderId}`} />
+                        <OrderData title="Order Number" value={`#${`${details.order.orderId}`.substring(0,6)}`} />
                         <Divider type="vertical" className="h-14 bg-secondryD" />
                         <OrderData
                           title="Buyer"
@@ -480,7 +418,7 @@ const SoldOrderDetails = ({ user, users }) => {
                         <Divider type="vertical" className="h-14 bg-secondryD" />
 
                         {
-                          status !== getStatus(1) || details.paymentSessionId !== "" ? <Col>
+                          status !== getStatus(1) ? <Col>
                             <Text className="block text-primaryC text-[13px] mb-2">
                               Status
                             </Text>
@@ -552,7 +490,7 @@ const SoldOrderDetails = ({ user, users }) => {
                       </Row>
                       <Row className="my-2 md:hidden flex-col gap-[6px] justify-between p-4 rounded">
                         <div className="flex gap-4">
-                          <NewOrderData className="w-2/4" title="Order Number" value={'#' + details.order.orderId} />
+                          <NewOrderData className="w-2/4" title="Order Number" value={'#' + `${details.order.orderId}`.substring(0,6)} />
                           <NewOrderData className="w-2/4" title="Buyer" value={details.order.purchasersCommonName} />
                         </div>
                         <div className="flex gap-4">
@@ -601,6 +539,7 @@ const SoldOrderDetails = ({ user, users }) => {
                         />
                       </div>
                     </Card>
+                    )}
                     {data?.length > 0 && data?.map((item) => {
                       return (
                         <ResponsiveOrderDetailCard data={item} />)
@@ -631,7 +570,6 @@ const SoldOrderDetails = ({ user, users }) => {
           />
 
         </div>
-      )}
       {message && openToastOrder("bottom")}
     </div>
   );