+++ marketplace/ui/src/contexts/users/actions.js
@@ -8,11 +8,12 @@ const actionDescriptors = {
 };
 
 const actions = {
-  fetchUsers: async (dispatch) => {
+  fetchUsers: async (dispatch, search) => {
     dispatch({ type: actionDescriptors.fetchUsers });
 
     try {
-      const response = await fetch(`${apiUrl}/users`, {
+      const query = search ? `?limit=10&queryFields=commonName&queryValue=${search}` : ''
+      const response = await fetch(`${apiUrl}/users${query}`, {
         method: HTTP_METHODS.GET,
       });