+++ strato/VM/EVM/evm-solidity/src/BlockApps/SolidVMStorageDecoder.hs
@@ -7,7 +7,7 @@
 module BlockApps.SolidVMStorageDecoder
   ( decodeSolidVMValues,
     decodeCacheValues,
-    decodeCacheValuesForMapping,
+    decodeCacheValuesForCollections,
     replayDeltas, -- Testing only
     ReplayFailure (..),
     synthesize, -- Testing only
@@ -51,20 +51,21 @@ bimapValue f (name', value') = do
 decodeCacheValues :: M.Map B.ByteString B.ByteString -> [(T.Text, Value)] -> [(T.Text, Value)]
 decodeCacheValues hxs prevState = either (error . (++ ": " ++ show hxs) . printf "SVM.decodeCacheValues: %s" . show) id $ do
   let parseM = bimapM (hexStorageToPath . HexStorage) (hexStorageToBasic . HexStorage)
-      isBasic (StoragePath (Field _ : _)) = True
+      isBasic (StoragePath ([Field _])) = True
+      isBasic (StoragePath [Field _, Field fieldBS]) = C8.unpack fieldBS /= "length"
       isBasic _ = False
   pathValues <- mapM parseM $ M.toList hxs
   let pathValues' = filter (isBasic . fst) pathValues
   finalState <- bimap show HM.toList $ case prevState of
     [] -> synthesize pathValues'
     tvs -> replayDeltas pathValues' . HM.fromList . map (first encodeUtf8) $ tvs
-
   mapM (bimapM bsToText return) finalState
 
-decodeCacheValuesForMapping :: M.Map B.ByteString B.ByteString -> [(T.Text, Value)]
-decodeCacheValuesForMapping hxs = either (error . (++ ": " ++ show hxs) . printf "SVM.decodeCacheValuesForMapping: %s" . show) id $ do
+decodeCacheValuesForCollections :: M.Map B.ByteString B.ByteString -> [(T.Text, Value)]
+decodeCacheValuesForCollections hxs = either (error . (++ ": " ++ show hxs) . printf "SVM.decodeCacheValuesForCollections: %s" . show) id $ do
   let parseM = bimapM (hexStorageToPath . HexStorage) (hexStorageToBasic . HexStorage)
       isBasic (StoragePath [Field _, MapIndex _]) = True
+      isBasic (StoragePath [Field _, ArrayIndex _]) = True
       isBasic _ = False
   pathValues <- mapM parseM $ M.toList hxs
   let pathValues' = filter (isBasic . fst) pathValues
@@ -84,6 +85,7 @@ valueToSolidityValue = \case
     ValueBytes _ b -> Just . SolidityValueAsString <$> (first show . decodeUtf8') b
     ValueBool b -> Right . Just $ SolidityBool b
     ValueInt _ _ n -> fromShowable n
+    ValueDecimal n -> Right . Just $ SolidityValueAsString $ decodeUtf8 n
     ValueAddress a -> fromShowable a
     ValueAccount a -> fromShowable a
     ValueString s -> fromText s
@@ -106,6 +108,7 @@ valueToSolidityValue = \case
     tshowIdx :: SimpleValue -> Either String T.Text
     tshowIdx = \case
       ValueInt _ _ n -> Right . T.pack . show $ n
+      ValueDecimal n -> Right . T.pack . show $ n
       ValueAddress a -> Right . T.pack . show $ a
       ValueAccount a -> Right . T.pack . show $ a
       ValueString t -> Right t
@@ -210,6 +213,7 @@ fromBasic = \case
   BBool b -> SimpleValue $ ValueBool b
   BInteger n -> SimpleValue $! valueInt n
   BString bs -> SimpleValue $! valueBytes bs
+  BDecimal v -> SimpleValue $! ValueDecimal v
   BAccount a -> SimpleValue $! ValueAccount a
   BContract _ c -> ValueContract c
   BEnumVal tipe name num -> ValueEnum (labelToText tipe) (labelToText name) (fromIntegral num)