+++ strato/VM/SolidVM/solid-vm/exec_src/ExprsNeeded.hs
@@ -84,6 +84,7 @@ expressionCrawler = \case
   BoolLiteral {} -> ["BoolLiteral"]
   HexaLiteral {} -> ["HexaLiteral"]
   NumberLiteral {} -> ["NumberLiteral"]
+  DecimalLiteral {} -> ["DecimalLiteral"]
   StringLiteral {} -> ["StringLiteral"]
   AccountLiteral {} -> ["AccountLiteral"]
   TupleExpression _ subexprs ->