+++ strato/core/blockstanbul/src/Blockchain/Blockstanbul/Authentication.hs
@@ -11,13 +11,8 @@ module Blockchain.Blockstanbul.Authentication
   )
 where
 
--- import Control.Monad (liftM2, liftM3, unless)
---import Control.Monad.Trans.Except
-
--- import Data.List (intercalate)
-
 import BlockApps.X509.Certificate
-import Blockchain.Blockstanbul.Messages
+import Blockchain.Blockstanbul.Messages hiding (sequence)
 import Blockchain.Blockstanbul.Model.Authentication
 import Blockchain.Blockstanbul.StateMachine
 import Blockchain.Data.ArbitraryInstances ()
@@ -30,20 +25,25 @@ import Blockchain.Strato.Model.Class (blockHash)
 import Blockchain.Strato.Model.ExtendedWord
 import Blockchain.Strato.Model.Keccak256
 import Blockchain.Strato.Model.Secp256k1
+import Blockchain.Strato.Model.Validator
 import Control.Applicative ((<|>))
 import Control.Lens as L
 import Control.Monad (liftM2, liftM3, unless)
 import qualified Control.Monad.Change.Alter as A
 import Control.Monad.Except
-import Control.Monad.Trans.Class (lift)
 import Data.Binary
+import Data.ByteString (ByteString)
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as BL
 import Data.Either.Extra
-import Data.Maybe (catMaybes, mapMaybe, fromJust)
+import Data.Maybe (catMaybes)
+import Data.List
+import Data.Maybe (fromJust)
+import Data.Set (Set)
 import qualified Data.Set as S
 import Test.QuickCheck
 import Text.Printf
+import Text.Format
 
 instance Arbitrary IstanbulExtra where
   arbitrary = liftM3 IstanbulExtra arbitrary arbitrary arbitrary
@@ -60,9 +60,6 @@ addValidators vs =
     uncookRawExtra
       . set istanbul (Just (IstanbulExtra vs Nothing []))
       . cookRawExtra
-
-getValidatorList :: Block -> ChainMembers
-getValidatorList x = L.view (istanbul . _Just . validatorList) (cookRawExtra $ L.view extraLens x)
 
 getProposerSeal :: Block -> Maybe Signature
 getProposerSeal x = do
@@ -129,10 +126,19 @@ commitmentSeal sha =
   let mesg = commitmentMessage sha
    in sign mesg
 
-verifyCommitmentSeal :: Keccak256 -> Signature -> Maybe Address
+verifyCommitmentSeal :: MonadError String m =>
+                        Keccak256 -> Signature -> m Address
 verifyCommitmentSeal sha sig =
   let mesg = commitmentMessage sha
-   in fromPublicKey <$> recoverPub sig mesg
+   in fromPublicKey <$> recoverPub' sig mesg
+
+recoverPub' :: MonadError String m =>
+               Signature -> ByteString -> m PublicKey
+recoverPub' sig v =
+  case recoverPub sig v of
+    Nothing -> throwError "can't recover public key"
+    Just val -> return val
+
 
 finalHash :: Block -> Keccak256
 finalHash =
@@ -142,12 +148,12 @@ finalHash =
     . over extraDataLens scrubCommitmentSeals
     . blockBlockData
 
-signBenfInfo :: (HasVault m) => (ChainMemberParsedSet, Bool, Int) -> m (Signature)
+signBenfInfo :: (HasVault m) => (Validator, Bool, Int) -> m (Signature)
 signBenfInfo bnf =
   let mesg = keccak256ToByteString $ hash $ BL.toStrict $ encode bnf
    in sign mesg
 
-verifyBenfInfo :: (ChainMemberParsedSet, Bool, Int) -> Signature -> Maybe Address
+verifyBenfInfo :: (Validator, Bool, Int) -> Signature -> Maybe Address
 verifyBenfInfo bnf sig =
   let mesg = keccak256ToByteString $ hash $ BL.toStrict $ encode (bnf)
    in fromPublicKey <$> recoverPub sig mesg
@@ -171,46 +177,89 @@ authenticate (IMsg (MsgAuth cm sig) tm) = do
   return (mAddress == cmAddress)
 authenticate _ = return True
 
-replayHistoricBlock :: (A.Selectable Address X509CertInfoState m) => ChainMembers -> Word256 -> Block -> m (Either String (Word256, ChainMemberParsedSet))
-replayHistoricBlock realValidators@(ChainMembers chainWorkAround) seqNo blk = runExceptT $ do
+
+getX509FromAddress' :: (MonadError String m, A.Selectable Address X509CertInfoState m) =>
+                       Address -> m X509CertInfoState
+getX509FromAddress' address = do
+  maybeCert <- getX509FromAddress address
+
+  case maybeCert of
+    Nothing ->
+--      error $ --will make this stricter once the full soln is in
+      throwError $
+        "Missing address in X509 certificate database: " ++ formatAddressWithoutColor address
+    Just v -> return v
+
+replayHistoricBlock :: (MonadError String m, A.Selectable Address X509CertInfoState m) =>
+                       Set Validator -> Word256 -> Block -> m (Word256, Validator)
+replayHistoricBlock realValidators seqNo blk = do
   let ExtraData {..} = cookRawExtra . L.view extraLens $ blk
   IstanbulExtra {..} <- liftEither $ maybeToEither "no istanbul metadata" _istanbul
   let mProp = verifyProposerSeal blk =<< _proposedSig
-      signers =
-        S.fromList
-          . mapMaybe (verifyCommitmentSeal (blockHash blk))
-          $ _commitment
       blockNo = fromIntegral . number . blockBlockData $ blk
-  -- signersX509 =
-  -- signersX509 <- liftEither $ maybeToEither "no istanbul metadata" mapMaybe =<<(getX509FromAddress (S.toList signers))
-  -- signersz <-  S.fromList $ (map getChainMemberFromX509 signersX509)
-  noAddress <- lift $ traverse getX509FromAddress (S.toList signers)
 
-  -- unless (all  (== True) (map isJust noAddress ))  $ liftEither $ Left $ printf "No address realated to singers"
+  signers <- sequence $ map (verifyCommitmentSeal (blockHash blk)) _commitment
+      
+  noAddress <- sequence $ map getX509FromAddress signers
+  
+  let signerRes = S.fromList $ ((chainMemberParsedSetToValidator . getChainMemberFromX509) <$> (catMaybes noAddress))
+  
+  -- Will make this stricter once full soln in
+  --noAddress <- sequence $ map getX509FromAddress' (signers :: [Address])
+  --let signerRes = S.fromList $ (chainMemberParsedSetToValidator . getChainMemberFromX509) <$>  noAddreqss
 
-  let signerRes = S.fromList $ ((getChainMemberFromX509) <$> (catMaybes noAddress))
+  
   unless (seqNo + 1 == blockNo) $
-    liftEither $ Left $ printf "unexpected block number: have %d, wanted %d" blockNo (seqNo + 1)
-  unless (realValidators == _validatorList) $
-    liftEither $ Left "mismatched validators"
-  prop <- liftEither $ maybeToEither "invalid proposer seal" mProp
-  --  maybe (Left "invalid proposer seal") Right mProp
+    throwError $ printf "unexpected block number: have %d, wanted %d" blockNo (seqNo + 1)
 
-  mPropCert <- lift $ (getX509FromAddress prop)
-  case mPropCert of
-    Nothing -> liftEither . Left $ "Proposer does not have an X509 certificate: " ++ formatAddressWithoutColor prop
-    Just propCert -> do
-      let propChainMember = getChainMemberFromX509 propCert
-      -- unless ( propT == Nothing) $ liftEither $ Left $ printf "No x509 "
-      -- propRes <-  getChainMemberFromX509  (getX509FromAddress prop)
-      unless (propChainMember `S.member` chainWorkAround) $
-        liftEither $ Left . printf "proposer %s not a validator" . formatAddressWithoutColor $ prop
-      unless (signerRes `S.isSubsetOf` chainWorkAround) $ do
-        --let unexplained = intercalate "," . map formatAddressWithoutColor . S.toList $ signerRes S.\\ chainWorkAround
-        liftEither $ Left $ "unknown signers: " -- ++ unexplained
-      unless (3 * S.size signerRes > 2 * S.size chainWorkAround) $
-        liftEither $ Left $ printf "not enough commit seals (have %d out of %d)" (S.size signerRes) (S.size chainWorkAround)
-      liftEither $ Right (fromIntegral $ seqNo + 1, propChainMember)
+  prop <- liftEither $ maybeToEither "invalid proposer seal" mProp
+
+  propCert <- getX509FromAddress' prop
+
+  let propChainMember = chainMemberParsedSetToValidator $ getChainMemberFromX509 propCert
+
+  unless (propChainMember `S.member` realValidators) $
+    throwError $
+--    error $
+      "proposer " ++ formatAddressWithoutColor prop ++ " (" ++ format propChainMember ++ ")  not a validator"
+      ++ "\nreal validator list: " ++ show (map format $ S.toList realValidators)
+
+  let expectedValidatorList = S.map chainMemberParsedSetToValidator $ unChainMembers _validatorList
+--  let expectedValidatorList = [c | CommonName _ _ c _ <- S.toList (unChainMembers _validatorList)]
+
+  unless (expectedValidatorList == realValidators) $
+    throwError $
+--    error $ --will make this stricter once the full soln is in
+      "real validator list doesn't match expected validator list for block #" ++ show (number . blockBlockData $ blk)
+      ++ "\nreal validator list: " ++ show (map format $ S.toList realValidators)
+      ++ "\nblock validator list: " ++ show (map format $ S.toList expectedValidatorList)
+        
+  unless (signerRes `S.isSubsetOf` realValidators) $ do
+        let unexplained = intercalate "," . map format . S.toList $ signerRes S.\\ realValidators
+        if (signerRes S.\\ realValidators) `S.isSubsetOf` futureValidatorsHack
+          then throwError $ "future validators " ++ show unexplained ++ " jumped the gun, signed block #" ++ show blockNo ++ " before they were authorized to do so.  I'll throw the block away and wait for another validator to send me the properly signed block"
+          else throwError $
+--          else error $  --will make this stricter once the full soln is in
+               "unknown signers in block #" ++ show blockNo ++ ": " ++ unexplained
+               ++ "\nsignerRes: " ++ show (map format $ S.toList signerRes)
+               ++ "\nreal validator list: " ++ show (map format $ S.toList realValidators)
+               ++ "\nblock validator list: " ++ show (map format $ S.toList expectedValidatorList)
+
+  unless (3 * S.size signerRes > 2 * S.size realValidators) $
+    throwError $
+--        error $ --will make this stricter once the full soln is in
+      printf "not enough commit seals (have %d out of %d)" (S.size signerRes) (S.size realValidators)
+      ++ ": signerRes = " ++ show signerRes
+      ++ ", realValidators = " ++ show realValidators
+        
+  return (fromIntegral $ seqNo + 1, propChainMember)
 
 isHistoricBlock :: Block -> Bool
 isHistoricBlock = (> 32) . B.length . L.view extraLens
+
+--Required hack to ignore known bugs in the production chain for now.  This should be removed once we restart the chain, or purge the bad blocks from the network
+futureValidatorsHack :: Set Validator
+futureValidatorsHack = S.fromList [
+  "service-account-io-stratomercata-dsnallapu",
+  "service-account-io-stratomercata-chessgm9" -- not only does this suffer from the future validator bug, but it also was a false x509 cert because the case of the commonName was wrong
+  ]