+++ strato/core/strato-p2p/src/Blockchain/EventException.hs
@@ -9,6 +9,7 @@ import Control.Exception.Lifted
 data EventException
   = PeerDisconnected
   | PeerNonResponsive
+  | CurrentlyTooManyPeers
   | EventBeforeHandshake Message
   | WrongGenesisBlock
   | NetworkIDMismatch