+++ strato/core/strato-redis-blockdb/src/Blockchain/Strato/RedisBlockDB.hs
@@ -114,6 +114,7 @@ import Blockchain.Strato.Model.ExtendedWord (Word256)
 import Blockchain.Strato.Model.Gas
 import Blockchain.Strato.Model.Keccak256
 import Blockchain.Strato.RedisBlockDB.Models as Models
+import Blockchain.Strato.Model.Validator (Validator(..))
 import Control.Arrow (second, (&&&), (***))
 import Control.Concurrent (threadDelay)
 import Control.Monad
@@ -230,10 +231,11 @@ putChainInfo cId cInfo = do
 isValidator ::
   CM.ChainMemberParsedSet ->
   Redis Bool
-isValidator val =
-  sismember (namespaceToKeyPrefix Validators) (toValue val) >>= \case
+isValidator (CM.CommonName _ _ v _) =
+  sismember (namespaceToKeyPrefix Validators) (toValue (Validator v)) >>= \case
     Right b -> pure b
     _ -> pure False
+isValidator _ = pure False
 
 getValidatorAddresses :: Redis [Address]
 getValidatorAddresses = do 
@@ -242,7 +244,7 @@ getValidatorAddresses = do
     Right keysBS -> (fmap userAddress . catMaybes) <$> (sequence $ (getCertFromParsedSet . fromValue) <$> keysBS)
 
 addValidators ::
-  [CM.ChainMemberParsedSet] ->
+  [Validator] ->
   Redis (Either Reply Status)
 addValidators [] = pure $ Right Ok
 addValidators vals =
@@ -251,7 +253,7 @@ addValidators vals =
     Left reply -> pure $ Left reply
 
 removeValidators ::
-  [CM.ChainMemberParsedSet] ->
+  [Validator] ->
   Redis (Either Reply Status)
 removeValidators [] = pure $ Right Ok
 removeValidators vals =